# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      password:
        maxLength: 255
        minLength: 6
        type: string
      username:
        maxLength: 255
        type: string
    required:
      - password
      - username
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      password:
        maxLength: 255
        minLength: 6
        type: string
      phone:
        maxLength: 20
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
      - email
      - first_name
      - last_name
      - password
      - username
    type: object
...
  common.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      error:
        $ref: "#/definitions/common.ResponseError"
      meta:
        $ref: "#/definitions/common.ResponseMeta"
    type: object
  common.ResponseError:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  common.ResponseMeta:
    properties:
      extra:
        additionalProperties: true
        type: object
      pagination:
        $ref: "#/definitions/common.PaginationMeta"
      timestamp:
        type: string
      trace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the KisanLink E-commerce service API.
  title: KisanLink E-commerce API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
        - application/json
      description: Authenticate user and return access token
      parameters:
        - description: Login credentials
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/auth.LoginRequest"
      produces:
        - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: "#/definitions/common.Response"
        "400":
          description: Invalid request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Invalid credentials
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: User Login
      tags:
        - Authentication
  /api/v1/auth/logout:
    post:
      consumes:
        - application/json
      description: Logout user and invalidate access token
      produces:
        - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: "#/definitions/common.Response"
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      security:
        - BearerAuth: []
      summary: User Logout
      tags:
        - Authentication
  /api/v1/auth/register:
    post:
      consumes:
        - application/json
      description: Register a new user account
      parameters:
        - description: User registration data
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/auth.RegisterRequest"
      produces:
        - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: "#/definitions/common.Response"
        "400":
          description: Invalid request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "409":
          description: User already exists
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: User Registration
      tags:
        - Authentication
  /api/v1/catalog:
    get:
      consumes:
        - application/json
      description: Retrieve a list of all catalog items with filtering and pagination
      parameters:
        - default: 1
          description: Page number
          in: query
          name: page
          type: integer
        - default: 20
          description: Items per page
          in: query
          name: limit
          type: integer
        - description: Filter by item type (PRODUCT, SERVICE, LABOUR)
          in: query
          name: item_type
          type: string
        - description: Filter by category
          in: query
          name: category
          type: string
        - description: Filter by subcategory
          in: query
          name: subcategory
          type: string
        - description: Filter by active status
          in: query
          name: is_active
          type: boolean
        - description: Filter by visibility (PRIVATE, ORG, NETWORK, PUBLIC)
          in: query
          name: visibility
          type: string
        - description: Minimum price filter
          in: query
          name: min_price
          type: number
        - description: Maximum price filter
          in: query
          name: max_price
          type: number
        - description: Search term
          in: query
          name: search
          type: string
        - collectionFormat: csv
          description: Filter by tags
          in: query
          items:
            type: string
          name: tags
          type: array
        - description: Filter by organization ID
          in: query
          name: organization_id
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    items:
                      $ref: "#/definitions/catalog.CatalogItemResponse"
                    type: array
                  meta:
                    allOf:
                      - $ref: "#/definitions/common.ResponseMeta"
                      - properties:
                          pagination:
                            $ref: "#/definitions/common.PaginationMeta"
                        type: object
                type: object
      summary: List all catalog items
      tags:
        - catalog
  /api/v1/catalog/{type}:
    get:
      consumes:
        - application/json
      description:
        Retrieve a list of catalog items filtered by type with additional
        filtering and pagination
      parameters:
        - description: Item type (products, services, labour)
          in: path
          name: type
          required: true
          type: string
        - default: 1
          description: Page number
          in: query
          name: page
          type: integer
        - default: 20
          description: Items per page
          in: query
          name: limit
          type: integer
        - description: Filter by category
          in: query
          name: category
          type: string
        - description: Filter by subcategory
          in: query
          name: subcategory
          type: string
        - description: Filter by active status
          in: query
          name: is_active
          type: boolean
        - description: Filter by visibility (PRIVATE, ORG, NETWORK, PUBLIC)
          in: query
          name: visibility
          type: string
        - description: Minimum price filter
          in: query
          name: min_price
          type: number
        - description: Maximum price filter
          in: query
          name: max_price
          type: number
        - description: Search term
          in: query
          name: search
          type: string
        - collectionFormat: csv
          description: Filter by tags
          in: query
          items:
            type: string
          name: tags
          type: array
        - description: Filter by organization ID
          in: query
          name: organization_id
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    items:
                      $ref: "#/definitions/catalog.CatalogItemResponse"
                    type: array
                  meta:
                    allOf:
                      - $ref: "#/definitions/common.ResponseMeta"
                      - properties:
                          pagination:
                            $ref: "#/definitions/common.PaginationMeta"
                        type: object
                type: object
      summary: List catalog items by type
      tags:
        - catalog
...
  /api/v1/catalog/{type}/{id}/price:
    put:
      consumes:
        - application/json
      description: Update the price of a catalog item with price history tracking
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Item type (products, services, labour, contracts)
          in: path
          name: type
          required: true
          type: string
        - description: Item ID
          in: path
          name: id
          required: true
          type: string
        - description: Price update data
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/catalog.PriceUpdateRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/catalog.CatalogItemResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "404":
          description: Not Found
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Update catalog item price
      tags:
        - Catalog Management
...
  /api/v1/catalog/{type}/{id}/unpublish:
    post:
      consumes:
        - application/json
      description: Unpublish a catalog item to make it inactive
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Item type (products, services, labour, contracts)
          in: path
          name: type
          required: true
          type: string
        - description: Item ID
          in: path
          name: id
          required: true
          type: string
        - description: Unpublish options
          in: body
          name: request
          schema:
            $ref: "#/definitions/catalog.UnpublishRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/catalog.CatalogItemResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "404":
          description: Not Found
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "409":
          description: Conflict
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Unpublish catalog item
      tags:
        - Catalog Management
...
  /status:
    get:
      description:
        Get comprehensive status dashboard with service health, dependencies,
        and alerts
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/health.StatusDashboard"
        "503":
          description: Service Unavailable
          schema:
            $ref: "#/definitions/health.StatusDashboard"
      summary: Get status dashboard
      tags:
        - health
swagger: "2.0"